requirements.txt
Flask
faiss-cpu
sentence-transformers
torch
groq
gunicorn